Do not install duplicated fluentd under bundler/gems/...
Closes: #263

In the previous versions, fluentd was installed twice.

  /opt/td-agent/lib/ruby/gems/2.7.0/bundler/gems/fluentd-793c9aa3a45b/
  /opt/td-agent/lib/ruby/gems/2.7.0/gems/fluentd-x.y.z

It is obviously redundant, but it is required to avoid "bundle exec
fluentd ...". This kind of redundancy is introduced by git source.

To fix the issue, use two approach:

* Fixate specific version by git_source (same as the previous way)
* Use pre-built gem and install it from local gem repository (new)

By using above approach, we can gain benefits (locking
Fluentd version as is and remove redundancy with fluentd installation)

NOTE: It may be simple to unify source FLUENTD_LOCAL_GEM_REPO but,
source ... syntax requires full path and it embeds full path into
Gemfile.lock, it is not acceptable because it vary in each build
environment, so it is better to use pre-built gem and use it (2-pass
approach)

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>