Ensure to install binstubs under /opt/td-agent/bin
In the previous version of TD Agent, binstubs are
not correctly installed for example, td command is missing
under /opt/td-agent/bin.

To fix this issue, gem pristine --only-executables
and --bindir option is used for generating binstubs correctly

NOTE: bundle install --binstubs or bundle binstubs is not appropriate
because it sets incorrect binstubs for BUNDLE_GEMFILE which is based
on staging directory path.

Signed-off-by: Kentaro Hayashi <kenhys@gmail.com>