apt: drop apt-key from test script
As apt-key will be removed in Debian 12 (bookworm), it is not
recommended to use apt-key. The replacement of apt-key is using
/etc/apt/trusted.gpg.d/ instead.

See
https://www.debian.org/releases/bullseye/amd64/release-notes/ch-information.en.html#obsolescense-and-deprecation

In this PR, we put a converted keyring under /usr/share/keyrings and
refer it by signed-by=... from /etc/apt/sources.list.d/confluent.list.
(/etc/apt/trusted.gpg.d is trusted by default in system wide, so it
may be better to limit a effective scope of keyring)

NOTE: GnuPG 2.x new keyring format (keybox) is not recognized, we need
to use legacy public ring v4.x for a moment. ( gnupg-ring: prefix in
--keyring argument does it explicitly)

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>