Use YAML.safe_load explicitly
If Psych backend v4.0.0 or later is used for YAML, it fails to load
because of incompatibility of YAML.load.

For example, it causes the following error.

  Psych::DisallowedClass: Tried to load unspecified class: Gem::Specification

As it is known that what class will be actually loaded, it should use
YAML.safe_load and permit class explicitly.

This fix is compatible with Psych v3.x and v4.x.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>