Add fluent-plugin-windows-exporter
Windows Exporter plugin is a system metrics plugin for Win32 platforms.
It can capture a wide range of metrics (CPU, memory, and network etc.),
and is very convenient for monitoring Windows nodes.

https://github.com/fluent-plugins-nursery/fluent-plugin-windows-exporter

The plugin enables Fluentd to act as a replacement of general system
monitoring solution (such as collectd or Prometheus), and should be
very useful for Windows administrators.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>