Fix cannot build fluentd correctly on Ubuntu bionic & focal
Due to CVE-2022-24765, Git 2.35.2 or later cannot run any command when
the paerent directory of a repository is owned by someone else.
To fix it, make sure to set owner of extracted source files &
directories to the current user (root).

`git config --add safe.directory /path/to/fluentd-source` can also avoid
this issue but it's not appropriate on self build case.

Signed-off-by: Takuro Ashie <ashie@clear-code.com>