yum: Ignore rpmlint error "E: env-script-interpreter"
Many gems include such shebang but it's expected for us since we use our
own ruby interpreter with additional `PATH`.

```
td-agent.x86_64: E: env-script-interpreter /opt/td-agent/lib/ruby/gems/3.1.0/gems/fluentd-1.14.6/bin/fluentd /usr/bin/env ruby
This script uses 'env' as an interpreter. For the rpm runtime dependency
detection to work, the shebang #!/usr/bin/env python  needs to be patched into
dependency on /usr/bin/env rather than the actual interpreter /usr/bin/python.
Alternatively, if the file should not be executed, then ensure that it is not
marked as executable or don't install it in a path that is reserved for
executables.
```

Signed-off-by: Takuro Ashie <ashie@clear-code.com>