migrate configuration files
In the previous versions, there are some configuration files and
log files are still kept under old path for compatibility.
It is better migrating to /opt/fluent completely, but we
need to consider upgrading from v4, too.

During upgrading from v4, seamless migration can be done by using
environment variable to specify the actual path because there is
no need to embed specific path to .service itself.
It aims not to be confused which configuration file should be applied.

Before:
 * Use old path for compatibility.
   * /etc/td-agent/td-agent.conf
   * /etc/td-agent/plugin
   * /var/log/td-agent/

After:

 * Use /opt/fluent as possible.
 * Keep /opt/td-agent as possible (only upgrading)
   * Moving content under /etc/td-agent to /etc/fluent/
   * Moving content under /var/log/td-agent to /var/log/fluent
   * Making symlink from /opt/td-agent /opt/fluent
   * Making symlink from /var/log/td-agent to /var/log/fluent
 * Note that /etc/logrotate.d/td-agent is not changed

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>