Add environment variables for builtin config (#878)
Add missing environment variables in
https://github.com/fluent/fluent-package-builder/pull/875

**NOTE**
We discussed about Windows support.
We need to prepare `Environment` in the registry during installation and
set the environment variables in order to work properly as a Windows
service.
Ref. [Set Environment variable for Windows
Service](https://community.usoft.com/share-your-best-practices-6/set-environment-variable-for-windows-service-2379)

We have decided that it does not support builtin config on Windows in
v6.0.0, since we do not have the time to implement and verify this.

---------

Signed-off-by: Shizuo Fujita <fujita@clear-code.com>