tmpfiles.d: ensure tmp dir for lock files not be cleaned up
We should exclude `/tmp/fluentd-lock-*` directory.
If `/tmp/fluentd-lock-*` is cleaned up while Fluentd is running:

* out_file can fail flushing.
* out_secondary can fail flushing.
* The supervisor process exit abnormally when Fluentd stops.

At least, RHEL 8 has the following default conf for tmpfiles.d:

/usr/lib/tmpfiles.d/tmp.conf

    q /tmp 1777 root root 10d

This can delete directories under `/tmp` if they are not used
for more than 10 days.

We don't need to protect the old lock files under
`/tmp/fluentd-lock-*/`, so use `X` instead of `x`.

In configs where lock files are not used, there is no problem
even if this directory is deleted.
In configurations that do use lock files, it’s unlikely that
this directory would not be updated for more than 10 days.
Therefore, the cases where this actually becomes an issue should
be quite limited.
If only out_secondary_file uses lock files, this can become an
issue.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>